Adicionado model para projeto
