+++ src/main/webapp/frontend/styles/shared-styles.html
@@ -71,13 +71,15 @@
         }
 
         .main-layout__title {
-            font-size: 1em;
+            font-size: 1.6em;
             margin: 0;
             /* Allow the nav-items to take all the space so they are centered */
             width: 0;
             line-height: 1;
             letter-spacing: -0.02em;
             font-weight: 500;
+            white-space: nowrap;
+            color: #2d83f4;
         }
 
         .main-layout > * {
